QA-74: Replace Jenkins' build trigger for GitLab ones
Two main changes:
- GitLab will not run Pipelines for PRs coming from forks. To workaround
this and continue with our current workflow, we will be creating
branches like pr_XXX for each incoming PR on a given microservice or
component.
- Modify the main triggerBuild logic to access GitLab API instead of
Jenkins. We use the same names for the build parameters. As per now the
Pipeline itself will do nothing, but the bot is ready for QA-75

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>