QA-107: Keep OS and Enterprise repos in sync
Note: This functionality relies on an external tool:
https://github.com/containous/gallienii to keep the repos in sync. When the
integration-test-runner detects a merge into one of the OS repositories with an
Enterprise fork, then an HTTP GET request is sent to the gallienii server, which
is running as a seperate service on the server. Then gallienii will run its sync
functionality which syncs all Enterprise repos with its OS counterpart. Note
that this is not specific to the merged PR, and hence all unsynchronized code
across all Enterprise repos will get a PR with the missing code. The gallienii
server will be listening to a configureable port, given by the GALLIENI_PORT
environment variable. Currently it syncs on all branches matching [0-9].[0-9].x
of master.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>