QA-107: Keep OS and Enterprise repos in sync
When the integration-test-runner detects a merge into one of the OS repositories
with an Enterprise fork, then a PR is created in the matching Enterprise
repository. This is done through locally fetching the branch in question on both
Enterprise and OS editions, and pushing them to a PR branch, by the name of
'mergeostoent_<pr-number> in the mender-test-bot repo. This sync will be done on
all release branches, or on master.

In case of merge conflicts, the bot will not merge, but rather create a PR in
the Enterprise repo which points at OS master. Then the PR branch must be
fetched locally, and merged with enterprise, for then pushing the merge commit
to the PR branch, prior to merging in Github.

This double merge is needed, because if the branches are not merged prior to
running the tests, the PR branch will not contain the Enterprise specific
functionality, and hence the tests will fail.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>