Fix: Build PR's moved out of the 'Draft' status
The Github PR's will not get the status 'opened', like a regular PR when moved
out of the 'Draft' state. However, it gets the status 'ready_for_review', which
currently is not a recognized action in the integration-test-runner.

This fix adds the 'ready_for_review' status to the accepted status for builds.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>