QA-192: Containerize tool and introduce Pipeline to the repo
Added the required files to build and run the tool through a Linux
container.

The docker-compose.yaml file is added aiming for local development while
the kubernetes.yaml file will describe the future deployment in the
production VM.

Added a pipeline to perform Go static checks and carry the build/publish
of the image.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>