QA-201: Vendor dependabot PR's
Since Github's Dependabot does not support vendoring in Go, a manual step has to
be applied after a PR is opened with a dependency upgrade.

This commit will add the functionality, by having the bot fetch the remote
branch, vendor and tidy the repository, and push the new commit.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>